Essentially ready for production.
Instructions for use: populate markdown files with content. Use
publications.markdown as a guide.

The schema of which files feed into which pages is defined by schema.py.
This probably does not need to be changed.

CSS can be changed as desired. New colors need to be chosen.

A few elements are missing from the CSS. Namely gravatar/icons/things
like that. The sidebar should be populated with easy access to other
me's on the internet --- with icons.